+++ internal/kibana/dashboard/lens_dashboard_app_validator_test.go
@@ -67,7 +67,6 @@ func Test_lensDashboardAppConfigModeValidator(t *testing.T) {
 			"hide_title":      types.BoolNull(),
 			"hide_border":     types.BoolNull(),
 			"drilldowns":      types.ListNull(drillElemType),
-			"drilldowns_json": jsontypes.NewNormalizedNull(),
 			"time_range":      tr,
 		})
 	}