+++ internal/kibana/dashboard/models_lens_dashboard_app_converters.go
@@ -431,8 +431,6 @@ func populateLensDashboardAppByReferenceFromAPI(
 		by.Drilldowns = nil
 	}
 
-	by.DrilldownsJSON = jsontypes.NewNormalizedNull()
-
 	pm.LensDashboardAppConfig = &lensDashboardAppConfigModel{
 		ByReference: &by,
 	}